requirements.txt
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0
requests>=2.31.0
urllib3>=2.1.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
dnspython>=2.4.0
pyjwt>=2.8.0
websocket-client>=1.6.0
reportlab>=4.0.0
Jinja2>=3.1.0
